Add limit to scatter and housekeeping
